% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{extract_metadata_df}
\alias{extract_metadata_df}
\title{Extract Metadata Data Frame}
\usage{
extract_metadata_df(project, include_non_experimental = FALSE)
}
\arguments{
\item{project}{A \code{dld8_project} object.}

\item{include_non_experimental}{Logical; include non-experimental files.}
}
\value{
Data frame of metadata (one row per file/chamber).
}
\description{
Extract Metadata Data Frame
}
