% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{get_marks_from_chamber}
\alias{get_marks_from_chamber}
\title{Extract Marks From a Chamber}
\usage{
get_marks_from_chamber(chamber_id, dld8_json)
}
\arguments{
\item{chamber_id}{Character: \code{"chamber_a"} or \code{"chamber_b"}.}

\item{dld8_json}{Parsed DLD8 list (from \code{read_dld8_json()}).}
}
\value{
List of marks.
}
\description{
Extract Marks From a Chamber
}
\keyword{internal}
