% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{new_dld8_project}
\alias{new_dld8_project}
\title{Construct a DLD8 Project Object}
\usage{
new_dld8_project(
  source_folder,
  rel_files,
  excluded_files,
  exclude_non_experimental,
  parse_errors = list()
)
}
\arguments{
\item{source_folder}{Root folder containing \code{.dld8} files.}

\item{rel_files}{Relative paths included in the project.}

\item{excluded_files}{Relative paths excluded in the project.}

\item{exclude_non_experimental}{Logical; whether non-experimental protocols were filtered.}

\item{parse_errors}{Optional list of parse errors.}
}
\value{
An object of class \code{dld8_project}.
}
\description{
Construct a DLD8 Project Object
}
\keyword{internal}
