% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{read_dld8_json}
\alias{read_dld8_json}
\title{Read a DLD8 File (BSON)}
\usage{
read_dld8_json(path)
}
\arguments{
\item{path}{Path to a \code{.dld8} file.}
}
\value{
List containing the first BSON document.
}
\description{
Reads a \code{.dld8} file into a list and removes problematic keys.
Requires \code{mongolite} (preferred) or \code{bson}.
}
\keyword{internal}
