% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{scan_dld8_files}
\alias{scan_dld8_files}
\title{Scan a Folder for \code{.dld8} Files}
\usage{
scan_dld8_files(source_folder)
}
\arguments{
\item{source_folder}{Folder to scan.}
}
\value{
Character vector of relative paths under \code{source_folder}.
}
\description{
Scan a Folder for \code{.dld8} Files
}
